PostGenerator/requirements.txt
api-display-purposes==0.0.3
APScheduler==3.6.3
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.1
clarifai==2.6.2
configparser==3.8.1
decorator==4.0.11
EasyProcess==1.1
emoji==2.0.0
filetype==1.0.5
future==0.18.2
googleapis-common-protos==1.56.4
googletrans==4.0.0rc1
grpcio==1.47.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.8.1
httpcore==0.9.1
httpx==0.13.3
huepy==1.2.1
hyperframe==5.2.0
idna==2.10
imageio==2.1.2
imageio-ffmpeg==0.4.7
iniconfig==1.1.1
instagram-private-api==1.6.0.0
instagram-private-api-extensions==0.3.8
instagrapi==1.16.29
instapy==0.6.16
Js2Py==0.71
jsonschema==2.6.0
MeaningCloud-python==2.0.0
mock==4.0.3
moviepy==0.2.3.2
numpy==1.23.2
outcome==1.2.0
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
plyer==2.0.0
proglog==0.1.10
protobuf==3.20.1
py==1.11.0
pycryptodomex==3.15.0
pydantic==1.9.2
pyjsparser==2.7.1
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
python-telegram-bot==13.13
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyVirtualDisplay==3.0
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
requests-toolbelt==0.9.1
responses==0.21.0
rfc3986==1.5.0
schedule==1.1.0
selenium==4.4.3
semantic-version==2.10.0
setuptools-rust==1.5.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tomli==2.0.1
tornado==6.2
tqdm==4.11.2
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
webdriverdownloader==1.1.0.3
wsproto==1.2.0
